[Feat/#74] 마이페이지 mypost myviewingparty api 연결

+++ app/src/main/java/umc/everyones/lck/data/dto/request/mypage/CancelHostViewingPartyMypageRequestDto.kt
@@ -1,5 +1,5 @@
 package umc.everyones.lck.data.dto.request.mypage
 
 data class CancelHostViewingPartyMypageRequestDto(
-    val viewingPartyId: Int
+    val viewingPartyId: Long
 )
