+++ app/src/main/java/umc/everyones/lck/data/dto/request/mypage/CancelParticipateViewingPartyMypageRequestDto.kt
@@ -1,5 +1,5 @@
 package umc.everyones.lck.data.dto.request.mypage
 
 data class CancelParticipateViewingPartyMypageRequestDto(
-    val viewingPartyId: Int
+    val viewingPartyId: Long
 )
